refactor: Namespace under `References::DefaultReference`
