+++ app/views/shared/_default_reference.haml
@@ -1,2 +1,2 @@
--default_reference = DefaultReference.get(session)
+-default_reference = References::DefaultReference.get(session)
 #default-reference{data: { id: default_reference.try(:id), keey: default_reference.try(:keey) }, style: 'display: none'}
